Feat: fusionner les hooks de formulaire
