+++ src/entities/core/hooks/index.ts
@@ -1,5 +1,4 @@
 "use client";
 
 export { default as useModelForm } from "./useModelForm";
-export { default as useModelForm2 } from "./useModelForm2";
 export type { FormMode, UseModelFormOptions, UseModelFormResult, FieldKey } from "./useModelForm";
