+++ src/entities/core/hooks/useCollectionForm.ts
@@ -1,6 +1,6 @@
 "use client";
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
-import useModelForm, { type FormMode } from "./useModelForm2";
+import useModelForm, { type FormMode } from "./useModelForm";
 
 type Id = string;
